% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring Manager.R
\name{add_keyring_connection}
\alias{add_keyring_connection}
\title{add a new user credential for a specific database}
\usage{
add_keyring_connection(
  user = "MATTHEW",
  database = "CCB",
  server = "jdbc:oracle:thin:@//sclzcisdbprod100.light.ci.seattle.wa.us:1557/RPTPRD",
  keyring_name = "MH",
  keyring_username = "HAMLINM_RO",
  keyring_service = "CCB",
  keyring_password = ""
)
}
\description{
add a new user credential for a specific database
}
