% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCB Queries.R
\name{ccb_customer_mapping}
\alias{ccb_customer_mapping}
\title{CCB Query that maps most of a customers relevant data and identifiers}
\usage{
ccb_customer_mapping(
  id = c(),
  is_badge_num = F,
  is_sa = F,
  is_acct_id = F,
  most_recent_rs = T
)
}
\arguments{
\item{id}{list of customer id's}

\item{is_badge_num}{logical: True if the ids provided are badge numbers}

\item{is_sa}{logical: True if the ids provided are service agreements}

\item{is_acct_id}{logical: True if the ids provided are account numbers}

\item{most_recent_rs}{logical: True if only interested in the most recent rate code for the customers}
}
\value{
Data frame of all identifiers mapped together
}
\description{
After establishing a ccb connection using the scl_connect() function, this
function given badge numbers, service agreements or account ids will map these ids
each other and attach additional ids for service point, service agreement status
and rate code
}
