% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring Manager.R
\name{delete_keyring_connection}
\alias{delete_keyring_connection}
\title{remove connection for a specific user and database}
\usage{
delete_keyring_connection(user = "MATTHEW", database = "CCB")
}
\description{
remove connection for a specific user and database
}
