% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - AVERAGE_RATES_FORECAST - Write
%   Functions.R
\name{fpustats_delete_average_rates_forecast}
\alias{fpustats_delete_average_rates_forecast}
\title{This function deletes average rates forecasts from the FPUSTATS Database}
\usage{
fpustats_delete_average_rates_forecast(
  forecast_vintage = 2026,
  eff_dt = "2026-02-12",
  connect = T,
  user = "MATTHEW",
  database = "EPMMART_RW"
)
}
\arguments{
\item{forecast_vintage}{Year of the vintage of interest}

\item{eff_dt}{upload date of the specific forecast you want to delete}

\item{connect}{T/F if you want the function to do the connect or not (typically false if looping over multiple reports)}

\item{user}{Keyring Manager username}

\item{database}{Keyring Manager database name}
}
\value{
Deletes specific forecasts from database and returns the print statement "Average rates forecast for vintage", forecast_vintage, "Uploaded on:", eff_dt,"has been deleted."
}
\description{
This function deletes average rates forecasts from the FPUSTATS Database
}
