% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - WEATHER_DATA - Write Functions.R
\name{fpustats_delete_weather_data}
\alias{fpustats_delete_weather_data}
\title{This function deletes weather data from the FPUSTATS Database}
\usage{
fpustats_delete_weather_data(year = 1980, month = 1, connect = T)
}
\arguments{
\item{year}{Year of the report of interest}

\item{month}{Month of the report of interest}

\item{connect}{T/F if you want the function to do the connect or not (typically false if looping over multiple reports)}
}
\value{
Deletes specific year/month from database and returns the print statement "Weather Data for", year,'-',month," has been deleted."
}
\description{
This function deletes weather data from the FPUSTATS Database
}
