% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - MSCS_HOURLY_KWH - Read Functions.R
\name{fpustats_read_mscs_hourly_data}
\alias{fpustats_read_mscs_hourly_data}
\title{This function writes a properly formatted Streetlight forecast to the FPUSTATS Database STREETLIGHT_FORECAST table}
\usage{
fpustats_read_mscs_hourly_data(
  start_date = "1980-01-01",
  end_date = "1980-01-15",
  read_all = F,
  connect = T
)
}
\arguments{
\item{start_date}{Year of the weather data of interest}

\item{end_date}{Month of the weather data of interest}

\item{read_all}{T if you want to pull down all weather data}

\item{connect}{True or false if you want the function to handle the FPUSTATS connect}
}
\value{
Writes forecast to FPUSTATS STREETLIGHT_FORECAST table and returns print statement "Successfully uploaded forecast vintage", unique(sl_forecast$FORECAST_VINTAGE)
}
\description{
This function writes a properly formatted Streetlight forecast to the FPUSTATS Database STREETLIGHT_FORECAST table
}
