% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - OPRV - Read Functions.R
\name{fpustats_read_oprv}
\alias{fpustats_read_oprv}
\title{This function reads oprv reports from the FPUSTATS Database}
\usage{
fpustats_read_oprv(
  year = 2024,
  month = 1,
  most_eff_dt = T,
  read_all = F,
  connect = T
)
}
\arguments{
\item{year}{Year of the report of interest}

\item{month}{Month of the report of interest}

\item{most_eff_dt}{T/F Do you want the most recently uploaded version of the report(s)}

\item{read_all}{T/F Do you want all oprv reports or not}
}
\value{
Returns all oprv records with the given parameter arguments
}
\description{
This function reads oprv reports from the FPUSTATS Database
}
