% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - SYSTEM_ADJ_RC_LOAD_DATA - Read
%   Functions.R
\name{fpustats_read_system_adj_rate_class_load}
\alias{fpustats_read_system_adj_rate_class_load}
\title{This function writes properly formatted hourly rate class load to the fpustats database}
\usage{
fpustats_read_system_adj_rate_class_load(
  start_date = "2024-01-01",
  end_date = "2024-01-15",
  read_all = F,
  connect = T
)
}
\arguments{
\item{start_date}{start date of interest}

\item{end_date}{end date of interest}

\item{read_all}{True if you want all available data}

\item{connect}{True if you want the function to handle the connection}
}
\value{
returns hourly rate class load data
}
\description{
This function writes properly formatted hourly rate class load to the fpustats database
}
