% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - WEATHER_DATA - Write Functions.R
\name{fpustats_update_weather_data}
\alias{fpustats_update_weather_data}
\title{This function updates the weather data in the FPUSTATS Database to the system date}
\usage{
fpustats_update_weather_data()
}
\value{
lets the user know if the weather data has been updated
}
\description{
This function updates the weather data in the FPUSTATS Database to the system date
}
