% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - AVERAGE_RATES_FORECAST - Write
%   Functions.R
\name{fpustats_write_average_rates_forecast}
\alias{fpustats_write_average_rates_forecast}
\title{This function writes a properly formatted average rates forecast to the FPUSTATS Database AVERAGE_RATES_FORECAST table}
\usage{
fpustats_write_average_rates_forecast(
  forecast_path =
    "I:/FINANCE/FPU/Sales and Revenue Actuals Model/Update Database Scripts/Average Rates/2026 Average Rates Forecast.xlsx",
  user = "MATTHEW",
  database = "EPMMART_RW"
)
}
\arguments{
\item{forecast_path}{Directory of the formatted vintage forecast path}

\item{user}{Keyring Manager username}

\item{database}{Keyring Manager database name}
}
\value{
Writes forecast to FPUSTATS AVERAGE_RATES_FORECAST table and returns print statement "Successfully uploaded forecast vintage", unique(avg_rates$FORECAST_VINTAGE)
}
\description{
This function writes a properly formatted average rates forecast to the FPUSTATS Database AVERAGE_RATES_FORECAST table
}
