% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - OPRV - Write Functions.R
\name{fpustats_write_oprv_report}
\alias{fpustats_write_oprv_report}
\title{This function writes a properly formatted oprv report to the FPUSTATS Database OPRV REPORT table}
\usage{
fpustats_write_oprv_report(
  report_path =
    "I:/FINANCE/FPU/Sales and Revenue Actuals Model/Update Database Scripts/OPRV/OPRV Files/2025-05 OPRV.xlsx",
  user = "MATTHEW",
  database = "EPMMART_RW",
  connect = T
)
}
\arguments{
\item{report_path}{Directory of the formatted oprv report}

\item{user}{Keyring Manager username}

\item{database}{Keyring Manager database name}
}
\value{
Writes report to FPUSTATS OPRV_REPORT table and returns print statement "Successfully uploaded report for", year,'-',month"
}
\description{
This function writes a properly formatted oprv report to the FPUSTATS Database OPRV REPORT table
}
