% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - STREETLIGHT_FORECAST - Write
%   Functions.R
\name{fpustats_write_streetlight_forecast}
\alias{fpustats_write_streetlight_forecast}
\title{This function writes a properly formatted Streetlight forecast to the FPUSTATS Database STREETLIGHT_FORECAST table}
\usage{
fpustats_write_streetlight_forecast(
  forecast_path =
    "I:/FINANCE/FPU/Sales and Revenue Actuals Model/Update Database Scripts/Streetlights/2026 Streetlight Forecast.xlsx",
  connect = T
)
}
\arguments{
\item{forecast_path}{Directory of the formatted vintage forecast path}

\item{user}{Keyring Manager username}

\item{database}{Keyring Manager database name}
}
\value{
Writes forecast to FPUSTATS STREETLIGHT_FORECAST table and returns print statement "Successfully uploaded forecast vintage", unique(sl_forecast$FORECAST_VINTAGE)
}
\description{
This function writes a properly formatted Streetlight forecast to the FPUSTATS Database STREETLIGHT_FORECAST table
}
