% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - SYSTEM_ADJ_RC_LOAD_DATA - Write
%   Functions.R
\name{fpustats_write_system_adj_rate_class_load}
\alias{fpustats_write_system_adj_rate_class_load}
\title{This function writes properly formatted hourly rate class load to the fpustats database}
\usage{
fpustats_write_system_adj_rate_class_load(
  hourly_rate_class_data = data.frame()
)
}
\arguments{
\item{hourly_rate_class_data}{cleaned and adjusted hourly rate class load}
}
\value{
Writes hourly data to FPUSTATS SYSTEM_ADJ_RC_LOAD_DATA table and returns print statement "Successfully uploaded data for ", year,'-',month"
}
\description{
This function writes properly formatted hourly rate class load to the fpustats database
}
