% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring Manager.R
\name{get_connect_creds}
\alias{get_connect_creds}
\title{get specific credentials for connection}
\usage{
get_connect_creds(user = "MATTHEW", database = "CCB")
}
\description{
get specific credentials for connection
}
