% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEN - Connect Functions.R
\name{scl_connect}
\alias{scl_connect}
\title{Generalized scl database connect function}
\usage{
scl_connect(database = "CCB")
}
\arguments{
\item{database}{Name of database user is trying to connect to}
}
\value{
Prints if connection was successful or not also will return parameters for db like MSCS
}
\description{
This function establishes R session connection to databases like
CCB, MSCS, ACCELA, and EPMMART
}
