% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEN - MSCS Queries.R
\name{scl_pull_badge_to_d1sp_map}
\alias{scl_pull_badge_to_d1sp_map}
\title{MSCS Query that maps badge number in CCB to d1_sp_id in MSCS and vice versa}
\usage{
scl_pull_badge_to_d1sp_map(ids = c("890903"), id_type = "BADGE_NBR")
}
\arguments{
\item{ids}{list of id's}

\item{id_type}{Name of the ID variable (BADGE_NBR or D1_SP_ID)}
}
\value{
Data frame of all identifiers mapped together
}
\description{
MSCS Query that maps badge number in CCB to d1_sp_id in MSCS and vice versa
}
