% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEN - CCB Queries.R
\name{scl_pull_customer_meter_meta}
\alias{scl_pull_customer_meter_meta}
\title{CCB Query that maps most of a customers relevant data and identifiers}
\usage{
scl_pull_customer_meter_meta(ids = c("0614615463"), id_type = "PREM_ID")
}
\arguments{
\item{ids}{list of customer id's}

\item{id_type}{Name of the ID variable}
}
\value{
Data frame of all identifiers mapped together
}
\description{
CCB Query that maps most of a customers relevant data and identifiers
}
