% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEN - MSCS Queries.R
\name{scl_pull_mscs_hourly_load}
\alias{scl_pull_mscs_hourly_load}
\title{MSCS Query that maps badge number in CCB to d1_sp_id in MSCS and vice versa}
\usage{
scl_pull_mscs_hourly_load(
  d1_sp_ids = c("746791443786", "950920972116"),
  start_date = "2025-01-01",
  end_date = "2025-01-08"
)
}
\arguments{
\item{d1_sp_ids}{list of d1_sp_id's}

\item{start_date}{start date of period for hourly data}

\item{end_date}{end date of period for hourly data}
}
\value{
Data frame of service point hourly data
}
\description{
MSCS Query that maps badge number in CCB to d1_sp_id in MSCS and vice versa
}
