% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEN - ODWP Queries.R
\name{scl_pull_system_load}
\alias{scl_pull_system_load}
\title{ODWP Query that pulls system load between two dates}
\usage{
scl_pull_system_load(start_date = "2024-03-01", end_date = "2024-04-01")
}
\arguments{
\item{start_date}{the start date of interest}

\item{end_date}{the end date of interest}
}
\value{
Data frame of system load between the two dates after correcting for the hour ending issue
}
\description{
ODWP Query that pulls system load between two dates
}
