% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEN - CCB Queries.R
\name{scl_rate_code_descr}
\alias{scl_rate_code_descr}
\title{CCB Query that maps rate codes to their full description}
\usage{
scl_rate_code_descr(connect = T)
}
\arguments{
\item{connect}{True if you want the function to make the connection to CCB}
}
\value{
Data frame with rate codes mapped to their full description
}
\description{
CCB Query that maps rate codes to their full description
}
