% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring Manager.R
\name{update_keyring_password}
\alias{update_keyring_password}
\title{update keyring password for a database}
\usage{
update_keyring_password(user = "MATTHEW", database = "CCB", password = "")
}
\description{
update keyring password for a database
}
