% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring Manager.R
\name{upload_backup_connections}
\alias{upload_backup_connections}
\title{get specific credentials for connection}
\usage{
upload_backup_connections(
  backup_dir =
    "I:/FINANCE/FPU/Matthew/Keyring Manager Backup/RSCL Keyring Manager Backup.xlsx"
)
}
\description{
get specific credentials for connection
}
