% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - OUL - Read Functions.R
\name{fpustats_read_oul_forecast}
\alias{fpustats_read_oul_forecast}
\title{This function reads oul forecasts from the FPUSTATS Database}
\usage{
fpustats_read_oul_forecast(
  forecast_vintage = 2026,
  year = 2024,
  most_eff_dt = T,
  read_all = F,
  user = "MATTHEW",
  database = "EPMMART_RW",
  connect = T
)
}
\arguments{
\item{forecast_vintage}{Which year's forecast are you interested in looking at}

\item{year}{Year within the forecast of interest}

\item{most_eff_dt}{T/F Do you want the most recently uploaded version of the forecast(s)}

\item{read_all}{T/F Do you want all oul forecasts or not}

\item{user}{Keyring Manager username}

\item{database}{Keyring Manager database name}
}
\value{
Returns all oul records with the given parameter arguments
}
\description{
This function reads oul forecasts from the FPUSTATS Database
}
