% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - WEATHER_DATA - Write Functions.R
\name{fpustats_update_weather_data}
\alias{fpustats_update_weather_data}
\title{This function writes a properly formatted Streetlight forecast to the FPUSTATS Database STREETLIGHT_FORECAST table}
\usage{
fpustats_update_weather_data(
  year = 2026,
  month = 1,
  user = "MATTHEW",
  database = "EPMMART_RW",
  connect = T
)
}
\arguments{
\item{year}{Year of the weather data of interest}

\item{month}{Month of the weather data of interest}

\item{user}{Keyring Manager username}

\item{database}{Keyring Manager database name}

\item{connect}{True or false if you want the function to handle the FPUSTATS connect}
}
\value{
Writes forecast to FPUSTATS STREETLIGHT_FORECAST table and returns print statement "Successfully uploaded forecast vintage", unique(sl_forecast$FORECAST_VINTAGE)
}
\description{
This function writes a properly formatted Streetlight forecast to the FPUSTATS Database STREETLIGHT_FORECAST table
}
