% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPUSTATS - OUL - Write Functions.R
\name{fpustats_write_oul_forecast}
\alias{fpustats_write_oul_forecast}
\title{This function writes a properly formatted OUL forecast to the FPUSTATS Database OUL_FORECAST table}
\usage{
fpustats_write_oul_forecast(
  forecast_path =
    "I:/FINANCE/FPU/Sales and Revenue Actuals Model/Update Database Scripts/OUL/2026 OUL Forecast.xlsx",
  user = "MATTHEW",
  database = "EPMMART_RW",
  connect = T
)
}
\arguments{
\item{forecast_path}{Directory of the formatted vintage forecast path}

\item{user}{Keyring Manager username}

\item{database}{Keyring Manager database name}
}
\value{
Writes forecast to FPUSTATS OUL_FORECAST table and returns print statement "Successfully uploaded forecast vintage", unique(oul_forecast$FORECAST_VINTAGE)
}
\description{
This function writes a properly formatted OUL forecast to the FPUSTATS Database OUL_FORECAST table
}
